+++ frontend/src/pages/Router.tsx
@@ -18,6 +18,7 @@ import SuManageDocTasksPage from "./SuManageDocTasks/SuManageDocTasksPage";
 import SuManageLogCRUDPage from "./SuManageLogCRUD/SuManageLogCRUDPage";
 import HomePage from "./HomePage/HomePage";
 import GaManageDocTasksPage from "./GaManageDocTasks/GaManageDocTasksPage";
+import GaGVDBsRetrParamsPage from "./GaGVDBsRetrParams/GaGVDBsRetrParamsPage";
 
 export const Router = createBrowserRouter([
   {
@@ -55,6 +56,10 @@ export const Router = createBrowserRouter([
           {
             path: "manage_queries",
             element: <GaManageDocTasksPage />,
+          },
+          {
+            path: "retrieval_parameters",
+            element: <GaGVDBsRetrParamsPage />,
           },
         ],
       },